+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.6.0] - 2020-07-30 - I was made for Refactorin' you baby 💋
+
 ### Changed
 
 - "Negate Expression" now generates simpler logical expressions, as you'd expect.
@@ -948,7 +950,8 @@ function doSomething(someData) {
 
 </details>
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/4.5.0...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/4.6.0...HEAD
+[4.6.0]: https://github.com/nicoespeon/abracadabra/compare/4.5.0...4.6.0
 [4.5.0]: https://github.com/nicoespeon/abracadabra/compare/4.4.0...4.5.0
 [4.4.0]: https://github.com/nicoespeon/abracadabra/compare/4.3.0...4.4.0
 [4.3.0]: https://github.com/nicoespeon/abracadabra/compare/4.2.0...4.3.0